// src/interfaces/prospect.interface.ts
export interface Prospect {
  prospect_id: number;
  activated: boolean;
  admission_type: string;
  admit_date: string;
  city: string;
  decision: string;
  email_address: string;
  ethnicity: string;
  first_name: string;
  last_name: string;
  level: string;
  middle_name: string;
  phone_number: string;
  primary_department_name: string;
  primary_program_name: string;
  secondary_department_name: string;
  secondary_program_name: string;
  sex: string;
  state: string;
  street_address_1: string;
  street_address_2: string;
  street_address_3: string;
  student_type: string;
  t_number: string;
  term: string;
  ualr_email: string;
  campaign_id: number;
  zip_code: string;
  created_at: string;
  updated_at: string;
}